backend/requirements.txt
# FastAPI and dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.10.3
redis>=4.5.4

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Utils
python-dotenv>=1.0.0
tenacity>=8.2.2
